stats updating to database
Added original male avatar.
Replaced filler avatar with original.
In CSS added a class and three ids for styling.
In HTML added some buttons to the player template.
In JS added extra stuff for each of the main buttons and added a few
extra click functions for the new buttons. Also modified the player
sidebar portion slightly. Was having some issues with the accounts-ui
portion. So the login stuff that would have been implemented in the
player sidebar were not implemented.